refactor: rename GroupPage file
